requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.0
blivet==3.4.4
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.4
cryptography==36.0.0
dasbus==1.6
distro==1.6.0
Django==4.1.3
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
humanize==3.13.1
idna==3.3
jsonschema==3.2.0
langtable==0.0.60
lxml==4.7.1
packaging==21.3
paramiko==2.11.0
pexpect==4.8.0
pid==2.2.3
ply==3.11
productmd==1.33
psycopg2-binary==2.9.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
pyenchant==3.2.2
pykickstart==3.36
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python-gettext==4.0
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
regex==2022.9.13
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
simpleline==1.9.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
texttable==1.6.2
urllib3==1.26.12
websocket-client==0.59.0
